reestructuring the api
